Add DateTime.Now Remove Environment.GetDateTime
