+++ System.Private.CoreLib/Pal/DateTime.CPM.cs
@@ -1,0 +1,10 @@
+﻿namespace System
+{
+    public partial struct DateTime
+    {
+        private static DateTime GetDateTime()
+        {
+            return new DateTime(0,0,0,0);
+        }
+    }
+}
