+++ System.Private.CoreLib/Pal/DateTime.Trs80.cs
@@ -1,0 +1,11 @@
+﻿using System.Runtime.InteropServices;
+
+namespace System
+{
+    public partial struct DateTime
+    {
+        [DllImport(Libraries.Runtime, EntryPoint = "GETDATETIME")]
+        private static extern DateTime GetDateTime();
+
+    }
+}
