+++ System.Private.CoreLib/Pal/Environment.Trs80.cs
@@ -18,9 +18,6 @@ namespace System
         /// </summary>
         public static int TickCount => GetTickCount() * 25;
 
-        [DllImport(Libraries.Runtime, EntryPoint = "GETDATETIME")]
-        public static extern DateTime GetDateTime();
-
         public static string NewLine => "\n";
     }
 }
