+++ System.Private.CoreLib/Pal/Environment.ZXSpectrum.cs
@@ -19,22 +19,6 @@ namespace System
         /// </summary>
         public static int TickCount => GetFrameCount() * 20;
 
-        //DllImport(Libraries.Runtime, EntryPoint = "GETDATETIME")]
-        public static DateTime GetDateTime()
-        {
-            var framesCounter = GetFramesCounter();
-            var seconds = (framesCounter / 50) % 60;
-            var minutes = (framesCounter / 3000) % 60;
-            var hours = (framesCounter / 180000) % 60;
-            var days = (framesCounter / 4320000) % 24;
-            return new DateTime(days, hours, minutes, seconds);
-        }
-
-        private static unsafe int GetFramesCounter()
-        {
-            return 65536 * *((byte*)23674) + 256 * *((byte*)23673) + *((byte*)23672);
-        }
-
         public static string NewLine => "\r";
     }
 }
